pyproject.toml
[project]
name = "geotech-llm"
version = "0.1.0"
description = "Advanced Geotechnical Engineering AI compliant agent"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "langchain",
    "langgraph",
    "chromadb",
    "unstructured",
    "streamlit",
    "python-dotenv",
    "langchain-community",
    "langchain-openai",
    "langchain-google-genai",
    "pandas",
    "numpy",
    "tiktoken",
    "pymupdf",
    "layoutparser",
    "torch", # Often needed for layoutparser/unstructured
    "torchvision"
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
